fix(shapes): unblock the lookup cache — rename the colliding ident, split the file
- `lookup` collided with `buffer/view.rs::lookup` in the root-holder gate's
  loose IDENT graph, which is module-blind past depth 0, so view.rs's body text
  (containing VIEW_REGISTRY) counted as reachable and the holder flipped to
  COVERED. Renamed to `lookup_ways`.
- `test_clear_shape_table` dropped every descriptor box while invalidating
  nothing, leaving dangling ways for any clear-then-lookup in tests.
- Extracted the cfg(test) helpers to a sibling file (2057 -> 1933 lines).
- Re-applied #8918's census-pinned carrier literal, which this branch's merge
  had reverted.
